+++ public/js/src/module/controller-webform.js
@@ -36,6 +36,11 @@ function init( formEl, data ) {
                 data.instanceStr = record.xml;
             }
 
+            if ( record && record.draft ) {
+                // Make sure that Enketo Core won't do the instanceID --> deprecatedID move
+                data.submitted = false;
+            }
+
             if ( data.instanceAttachments ) {
                 fileManager.setInstanceAttachments( data.instanceAttachments );
             }