+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- If an auto-saved record is recovered, a deprecatedID is generated even though the record has never been submitted.
+
 [2.3.0] - 2020-05-01
 ---------------------
 ##### Added